packages/x402-larkin-py/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "larkinsh-x402"
version = "0.1.0"
description = "Python SDK for Larkin — wallet authorization for x402-paid APIs."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["x402", "larkin", "agent", "wallet", "authorization"]
authors = [{ name = "Larkin" }]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "httpx>=0.27",
  "pynacl>=1.5",
]

[project.urls]
Homepage = "https://larkin.sh"
Repository = "https://github.com/larkin-dev/larkin.git"
"Bug Tracker" = "https://github.com/larkin-dev/larkin/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/larkinsh_x402"]
